requirements.txt
pillow==11.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
setuptools==58.1.0